Extract logic from `FieldSelect` which determines if field is qualified.
