+++ graylog2-web-interface/src/views/components/aggregationwizard/FieldSelect.tsx
@@ -20,12 +20,9 @@ import * as Immutable from 'immutable';
 import styled, { css } from 'styled-components';
 import type { SelectInstance, GroupBase } from 'react-select';
 
-import type { FieldTypeCategory } from 'views/logic/aggregationbuilder/Pivot';
-import { DateType, ValuesType } from 'views/logic/aggregationbuilder/Pivot';
 import { defaultCompare } from 'logic/DefaultCompare';
 import FieldTypesContext from 'views/components/contexts/FieldTypesContext';
 import Select from 'components/common/Select';
-import type { Property } from 'views/logic/fieldtypes/FieldType';
 import type FieldTypeMapping from 'views/logic/fieldtypes/FieldTypeMapping';
 import FieldTypeIcon from 'views/components/sidebar/fields/FieldTypeIcon';
 import type FieldType from 'views/logic/fieldtypes/FieldType';
@@ -44,6 +41,7 @@ type Props = {
   clearable?: boolean,
   excludedFields?: Array<string>,
   id: string,
+  isFieldQualified?: (field: FieldTypeMapping) => boolean,
   menuPortalTarget?: HTMLElement,
   name: string,
   onChange: (fieldName: string) => void,
@@ -51,36 +49,12 @@ type Props = {
   openMenuOnFocus?: boolean,
   persistSelection?: boolean,
   placeholder?: string,
-  properties?: Array<Property>,
-  qualifiedTypeCategory?: FieldTypeCategory,
   selectRef?: React.Ref<SelectInstance<unknown, boolean, GroupBase<unknown>>>,
   size?: 'normal' | 'small',
   value: string | undefined,
 }
 
 const sortByLabel = ({ label: label1 }: { label: string }, { label: label2 }: { label: string }) => defaultCompare(label1, label2);
-
-const hasProperty = (fieldType: FieldTypeMapping, properties: Array<Property>) => {
-  const fieldProperties = fieldType?.type?.properties ?? Immutable.Set();
-
-  return properties
-    .map((property) => fieldProperties.contains(property))
-    .find((result) => result === false) === undefined;
-};
-
-const isFieldQualified = (field: FieldTypeMapping, properties: Array<Property>, qualifiedTypeCategory: FieldTypeCategory | undefined) => {
-  if (properties) {
-    return hasProperty(field, properties);
-  }
-
-  if (qualifiedTypeCategory) {
-    const fieldTypeCategory = field.type.type === 'date' ? DateType : ValuesType;
-
-    return qualifiedTypeCategory === fieldTypeCategory;
-  }
-
-  return true;
-};
 
 const UnqualifiedOption = styled.span(({ theme }) => css`
   color: ${theme.colors.variant.light.default};
@@ -105,6 +79,7 @@ const FieldSelect = ({
   clearable,
   excludedFields,
   id,
+  isFieldQualified,
   menuPortalTarget,
   name,
   onChange,
@@ -112,8 +87,6 @@ const FieldSelect = ({
   openMenuOnFocus,
   persistSelection,
   placeholder,
-  properties,
-  qualifiedTypeCategory,
   selectRef,
   size,
   value,
@@ -127,10 +100,10 @@ const FieldSelect = ({
       label: field.name,
       value: field.name,
       type: field.type,
-      qualified: isFieldQualified(field, properties, qualifiedTypeCategory),
+      qualified: isFieldQualified(field),
     }))
     .toArray()
-    .sort(sortByLabel), [activeQuery, excludedFields, fieldTypes.queryFields, properties, qualifiedTypeCategory]);
+    .sort(sortByLabel), [activeQuery, excludedFields, fieldTypes.queryFields, isFieldQualified]);
 
   return (
     <Select options={fieldOptions}
@@ -141,6 +114,7 @@ const FieldSelect = ({
             openMenuOnFocus={openMenuOnFocus}
             persistSelection={persistSelection}
             clearable={clearable}
+            inputProps={{ 'aria-label': placeholder }}
             placeholder={placeholder}
             name={name}
             value={value}
@@ -159,13 +133,12 @@ FieldSelect.defaultProps = {
   autoFocus: undefined,
   className: undefined,
   clearable: false,
-  qualifiedTypeCategory: undefined,
+  isFieldQualified: () => true,
   excludedFields: [],
   onMenuClose: undefined,
   openMenuOnFocus: undefined,
   persistSelection: undefined,
   placeholder: undefined,
-  properties: undefined,
   selectRef: undefined,
   size: 'small',
   menuPortalTarget: undefined,
